requirements.txt
tabula-py==2.3.0
# csv